python-SkillMatchAI/requirements.txt
Flask>=2.2
flask-cors
python-dotenv
google-genai
gunicorn
